+++ app/controllers/StockTransactionController.ts
@@ -1,5 +1,5 @@
 import * as DB from '../../db';
-import * as Types from '../../src/lib/types';
+import * as Types from '../../lib/types';
 
 export class StockTransactionController {
