+++ db/portfolio/Portfolio.ts
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose';
-import * as Types from '../../src/lib/types';
+import * as Types from '../../lib/types';
 import { ITransactionDocument } from '../transaction';
 
 // Interface
