+++ server/api/index.ts
@@ -1,5 +1,4 @@
 import express from 'express';
-import { Logger, accessControlAllowOrigin } from './util';
 import portfoliosApi from './portfolios-api';
 import accountActivityApi from './account-activity-api';
 import portfolioSummaryApi from './portfolio-summary-api';
