--- src/index.ts
@@ -1,32 +1,0 @@
-import bodyParser from 'body-parser';
-import express from 'express';
-import path from 'path';
-import api from './api';
-import { connectDb } from '../db';
-
-const app = express();
-
-app.set('port', process.env.PORT);
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'pug');
-
-app.use(function (req, res, next) {
-  res.header('Access-Control-Allow-Origin', '*');
-  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
-  next();
-});
-
-app.use(express.urlencoded({ extended: true })); // for parsing application/x-www-form-urlencoded
-
-// Database
-connectDb(process.env.MONGODB_URL);
-
-// Serve Static files
-app.use('/public', express.static(path.join(__dirname, 'public')));
-
-// Api
-app.use('/api', api);
-
-// Listen
-app.listen(app.get('port'));
-console.log(`Server started. Port: ${app.get('port')}`);
