--- src/views/holdings.pug
@@ -1,21 +1,0 @@
-extends ./layout/default.pug
-
-//- Mixins
-include ./mixins/table
-include ./mixins/form
-
-block body
-    .container
-        h2 Holdings
-        .row
-            .col-8
-                +table(tableColumns, data, getDeleteUrl)
-            .col-4
-                h3 New Portfolio
-                form(action="/holdings", method="post")
-
-                    +formGroup('Symbol', 'symbol')
-                    +formGroup('Avg. Cost', 'avgCost')
-                    +formGroup('Quantity', 'quantity')
-
-                    button(type="submit" class="btn btn-primary") Submit
