--- src/views/index.pug
@@ -1,7 +1,0 @@
-extends ./layout/default.pug
-
-//- Mixins
-include ./mixins/table
-
-block body
-    h2 Home
