--- src/views/layout/default.pug
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-html(lang="en")
-  head
-    meta(charset="UTF-8")
-    meta(name="viewport", content="width=device-width, initial-scale=1.0")
-    meta(http-equiv="X-UA-Compatible", content="ie=edge")
-    title Stocks
-    link(rel="stylesheet" type="text/css" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous")
-    //- link(rel="stylesheet" type="text/css" href="/public/styles/main.css")
-    block head
-  body
-    .container
-      //- nav
-      //-   ul.nav
-      //-     li.nav-item
-      //-       a.nav-link(href="/") Home
-      //-     li.nav-item
-      //-       a.nav-link(href="/portfolios") Portfolios
-      block body
