--- src/views/mixins/form.pug
@@ -1,4 +1,0 @@
-mixin formGroup(label, key)
-    .form-group
-        label(for=`${key}`)= label
-        input(type="text" class="form-control" id=`${key}` name=`${key}`)
