--- src/views/mixins/table.pug
@@ -1,19 +1,0 @@
-mixin table(tableColumns, data)
-    if tableColumns && tableColumns.length
-        table.table.table-bordered
-            thead
-                tr
-                    each col in tableColumns
-                        - var className = col.key + ' ' + col.format
-                        th(class=className)= col.label
-            tbody
-                if data && data.length
-                    each d, i in data
-                        tr(id=`${d._id}`)
-                            each col in tableColumns
-                                - var className = col.key + ' ' + col.format
-                                - var value = d[col.key]
-                                td(class=className)= col.formatFn(value)
-                else
-                    tr
-                        td(colspan=`${tableColumns.length}` style="text-align:center;background:#ddd") No data
