--- src/views/portfolios.pug
@@ -1,19 +1,0 @@
-extends ./layout/default.pug
-
-include ./mixins/table
-include ./mixins/form
-
-block body
-    .container
-        h2 Portfolios
-        .row
-            .col-8
-                +table(tableColumns, data, getDeleteUrl)
-            .col-4
-                form(action="/holdings", method="post")
-
-                    +formGroup('Symbol', 'symbol')
-                    +formGroup('Avg. Cost', 'avgCost')
-                    +formGroup('Quantity', 'quantity')
-
-                    button(type="submit" class="btn btn-primary") Submit
