--- src/views/quotes.pug
@@ -1,8 +1,0 @@
-extends ./layout/default.pug
-
-//- Mixins
-include ./mixins/table
-
-block body
-    h2 Quotes
-    +table(tableColumns, data)
